arrumando bugs

+++ backend/src/database/services/orderItensService.js
@@ -37,11 +37,50 @@ export const listItensOrdersS = (pedidoId) => {
 
 export const deleteItensOrdersS = (id) => {
     try {
+        const getPedidoIdSql = `SELECT pedidoId FROM itens_order WHERE id = ?`;
+        const result = db.prepare(getPedidoIdSql).get(id);
+
+        if (!result) {
+            console.log('Item não encontrado para exclusão');
+            return false;
+        }
+
+        const { pedidoId } = result;
+
         const sql = 'DELETE FROM itens_order WHERE id = ?';
         db.prepare(sql).run(id);
+
+        updateValueOrderS(pedidoId);
+
         return true;
     } catch (error) {
         console.log("Erro ao deletar item de pedido:", error.message);
         return false;
     }
 }
+
+export const updateValueOrderS = (pedidoId) => {
+    try {
+        const sql = `
+            SELECT SUM(produtoValor) AS total
+            FROM itens_order
+            WHERE pedidoId = ?
+        `;
+
+        const result = db.prepare(sql).get(pedidoId);
+        const totalPedido = result ? result.total : 0;
+
+        const updateSql = `
+            UPDATE orders
+            SET total = ?
+            WHERE id = ?
+        `;
+
+        db.prepare(updateSql).run(totalPedido, pedidoId);
+
+        return totalPedido;  // Retorna o valor atualizado
+    } catch (error) {
+        console.error("Erro ao atualizar valor do pedido:", error.message);
+        throw new Error('Erro ao atualizar valor do pedido');
+    }
+};
